Port Rhino code to the Cycles 5.2 APIs
All 13 Cycles libraries now compile CPU-only; ccycles is down from 213 to
92 errors.

SVM compiler and kernel: 5.2 replaced the packed add_node(type, a, b, c) /
encode_uchar4 encoding with per-node typed structs. Rather than rewrite
Rhino's 23 SVM nodes and their kernel-side readers, the original wire format
is kept behind a compatibility layer - add_node_packed / encode_uchar4 /
stack_assign_if_linked / input_link(ShaderInput *) on SVMCompiler, and
read_node / read_node_float / svm_unpack_node_uchar2 / svm_unpack_node_uchar4 /
svm_rhino_read_packed in kernel/svm/util.h. The 23 RHINO_NODE cases reassemble
the uint4 the 5.2 interpreter no longer reads up front.

Other API changes absorbed:
 - integrator_state_read_ray / write_ray lost their KernelGlobals argument
 - PATH_RAY_CAMERA renamed to PATH_RAY_VISIBILITY_CAMERA
 - primitive_surface_attribute_float2/3 became templated, without derivatives
 - Attribute::data_float2/3 became data_for_write<T>
 - Mesh::reserve_mesh became resize_mesh; vertices moved into the
   ATTR_STD_POSITION attribute as packed_float3
 - NodeType::types() replaced by type_names() plus find()
 - DisplayDriver::clear became zero, graphics_interop_get became
   graphics_interop_get_device
 - Scene, ShaderGraph and Node are unique_ptr-owned; Shader::set_graph takes
   unique_ptr&&, ShaderGraph::add became add_node
 - sin/cos/acos/atan2/fmod are float3-only in 5.2, so scalar call sites use
   the f-suffixed forms
 - ASSERT came from OpenImageIO's dassert.h, removed in OIIO 3.x; replaced
   with standard assert

KERNEL_FEATURE_CLIPPING_PLANES shares the OSL camera bit: all 32 feature bits
are assigned upstream and Rhino builds without OSL. If OSL is ever enabled,
kernel_features must widen and this needs a bit of its own.

+++ src/kernel/svm/image.h
@@ -52,7 +52,7 @@ ccl_device_inline float3 env_spherical(float3 R)
     phi = (z >= 0.0f ? 0.5f * M_PI_F : -0.5f * M_PI_F);
   }
   else {
-    theta = atan2(y, x);
+    theta = atan2f(y, x);
     if (theta < 0.0f)
       theta += 2.0f * M_PI_F;
 
@@ -100,7 +100,7 @@ ccl_device_inline float3 env_emap_act(float3 R)
   if (fDivisor < FLT_MIN)
     fDivisor = FLT_MIN;
 
-  float f = sin(0.5f * acos(z)) / fDivisor;
+  float f = sinf(0.5f * acosf(z)) / fDivisor;
 
   float px = -x * f;
   float py = y * f;
@@ -131,7 +131,7 @@ ccl_device_inline float3 env_light_probe(float3 R)
   if (fDivisor < FLT_MIN)
     fDivisor = FLT_MIN;
 
-  float f = (acos(z) / M_PI_F) / fDivisor;
+  float f = (acosf(z) / M_PI_F) / fDivisor;
 
   float px = x * f;
   float py = y * f;
