+++ src/kernel/svm/svm_rhino_procedurals.h
@@ -1041,7 +1041,7 @@ ccl_device float4 waves_texture(float3 uvw,
   float value = 0.0f;
   if (hillParameter > hillMiddle - 0.5f * hillLength) {
     if (hillParameter < hillMiddle + 0.5f * hillLength) {
-      value = 0.5f + 0.5f * sin((hillParameter - hillMiddle) / (hillLength)*PI);
+      value = 0.5f + 0.5f * sinf((hillParameter - hillMiddle) / (hillLength)*PI);
     }
     else
       value = 1.0f;
@@ -1778,7 +1778,7 @@ ccl_device float3 light_probe_to_world(float2 uv)
   float v = uv.y;
 
   float d = sqrt((0.5f - u) * (0.5f - u) + (0.5f - v) * (0.5f - v)) * 2.0f;
-  float cosine = cos(PI * d);
+  float cosine = cosf(PI * d);
   float factor = 2.0f * sqrt(1.0f - cosine * cosine);
 
   float3 vec;
@@ -1797,7 +1797,7 @@ ccl_device float2 world_to_lightprobe(float3 n)
 
   float fDivisor = sqrt(x * x + y * y);
 
-  float f = (acos(z) / PI) / fDivisor;
+  float f = (acosf(z) / PI) / fDivisor;
 
   float px = x * f;
   float py = y * f;
@@ -1813,12 +1813,12 @@ ccl_device float3 equirect_to_world(float2 uv)
   float theta = (uv.x - 0.5f) * TWO_PI;
   float phi = -(uv.y - 0.5f) * PI;
 
-  float cosphi = cos(phi);
+  float cosphi = cosf(phi);
 
   float3 vec;
-  vec.x = sin(theta) * cosphi;
-  vec.y = sin(phi);
-  vec.z = cos(theta) * cosphi;
+  vec.x = sinf(theta) * cosphi;
+  vec.y = sinf(phi);
+  vec.z = cosf(theta) * cosphi;
 
   return vec;
 }
@@ -1836,7 +1836,7 @@ ccl_device float2 world_to_equirect(float3 n)
     phi = (z >= 0.0f ? 0.5f * PI : -0.5f * PI);
   }
   else {
-    theta = atan2(y, x);
+    theta = atan2f(y, x);
     if (theta < 0.0f)
       theta += 2.0f * PI;
 
@@ -2249,12 +2249,12 @@ ccl_device float3 emap_to_world(float2 uv)
   r = min(r, 1.0f);
 
   float t = 2.0f * asin(r);
-  float s = sin(t) / r;
+  float s = sinf(t) / r;
 
   float3 vec;
   vec.x = -s * x;
   vec.y = s * y;
-  vec.z = cos(t);
+  vec.z = cosf(t);
 
   return vec;
 }
@@ -2291,8 +2291,8 @@ ccl_device float2 world_to_hemispherical(float3 n)
 
 ccl_device Transform rotation_matrix(float3 axis, float angle)
 {
-  float s = -sin(angle);
-  float c = cos(angle);
+  float s = -sinf(angle);
+  float c = cosf(angle);
   float oc = 1.0f - c;
 
   Transform tf;
@@ -3554,7 +3554,7 @@ ccl_device void ProcessDotColor(KernelGlobals kg,
     s = hsb.y;
     b = hsb.z;
 
-    float shiftedHue = fmod(h + hueAdjust, 360.0f);
+    float shiftedHue = fmodf(h + hueAdjust, 360.0f);
 
     if (shiftedHue < 0)
       shiftedHue += 360.0f;
